backend/requirements/base.txt
Django>=5.0,<5.1
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=23.5,<24.0
djangorestframework-simplejwt>=5.3,<6.0
dj-rest-auth>=5.0,<6.0
django-allauth>=0.61,<1.0
channels>=4.0,<5.0
channels-redis>=4.1,<5.0
daphne>=4.0,<5.0
celery>=5.3,<6.0
redis>=5.0,<6.0
psycopg2-binary>=2.9,<3.0
Pillow>=10.2,<11.0
python-decouple>=3.8,<4.0
gunicorn>=21.2,<22.0
whitenoise>=6.6,<7.0
